Move ModalForm to own component
