+++ src/App.js
@@ -4,7 +4,7 @@ import styled from 'styled-components/macro'
 import CategoryList from './components/Card/CategoryList'
 import categories from './components/common/categories'
 import InitialData from './components/common/InitialData'
-import Form from './components/Form/Form'
+import ModalForm from './components/Form/ModalForm'
 import logo from './components/img/logo.png'
 import TimeRemaining from './components/TimeRemaining'
 import { loadFromLocal, saveToLocal } from './utils'
@@ -37,34 +37,24 @@ function App() {
           categories={categories}
           setActivities={setActivities}
           deleteActivity={deleteActivity}
+          setIsopen={setIsOpen}
         />
       </ScrollContainer>
 
       <Footer></Footer>
 
-      <Modal
-        isOpen={modalIsOpen}
-        onRequestClose={closeModal}
-        style={customStyles}
-      >
-        <CloseModal onClick={closeModal}>&times;</CloseModal>
-        <Form categories={categories} onSubmit={addActivity} />
-      </Modal>
+      <ModalForm
+        modalIsOpen={modalIsOpen}
+        categories={categories}
+        activities={activities}
+        setActivities={setActivities}
+        setIsOpen={setIsOpen}
+      />
     </AppGrid>
   )
 
   function openModal() {
     setIsOpen(true)
-  }
-
-  function closeModal() {
-    setIsOpen(false)
-  }
-
-  function addActivity(activity) {
-    const newActivities = [activity, ...activities]
-    setActivities(newActivities)
-    saveToLocal('activities', newActivities)
   }
 
   function deleteActivity(activity) {
@@ -123,33 +113,6 @@ const OpenModal = styled.button`
   display: block;
   margin: auto;
 `
-
-const CloseModal = styled.button`
-  border: none;
-  background: #1a1919;
-  color: grey;
-  font-size: 18px;
-  position: absolute;
-  left: 88%;
-  top: 2%;
-
-  :hover {
-    color: white;
-  }
-`
-
-const customStyles = {
-  content: {
-    top: '95%',
-    left: '50%',
-    right: 'auto',
-    bottom: 'auto',
-    marginRight: '-50%',
-    transform: 'translate(-50%, -50%)',
-    border: 'none',
-    background: '#1a1919',
-  },
-}
 
 const AddStyled = styled.div`
   border: 3px solid goldenrod;
